Moving square folder to a components
