+++ React/tic-tac/src/app/screens/game/components/board/components/Square/style.scss
